+++ packages/ui-components/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@unstoppabledomains/ui-components",
-  "version": "0.0.51",
+  "version": "0.0.51-browser-extension.1",
   "private": true,
   "description": "An open and reusable suite of Unstoppable Domains management components",
   "keywords": [
