+++ packages/ui-components/src/components/Chat/modal/ChatModal.tsx
@@ -37,15 +37,15 @@ import config from '@unstoppabledomains/config';
 import {makeStyles} from '@unstoppabledomains/ui-kit/styles';
 
 import {useFeatureFlags} from '../../../actions/featureFlagActions';
-import {
-  getDomainSignatureExpiryKey,
-  getDomainSignatureValueKey,
-} from '../../../components/Wallet/ProfileManager';
 import useFetchNotifications from '../../../hooks/useFetchNotification';
 import {fetchApi, isDomainValidForManagement} from '../../../lib';
 import {notifyEvent} from '../../../lib/error';
 import useTranslationContext from '../../../lib/i18n';
 import type {SerializedCryptoWalletBadge} from '../../../lib/types/badge';
+import {
+  getDomainSignatureExpiryKey,
+  getDomainSignatureValueKey,
+} from '../../../lib/types/domain';
 import type {
   SerializedRecommendation,
   SerializedUserDomainProfileData,
