+++ packages/ui-components/src/components/Chat/protocol/upload.ts
@@ -2,13 +2,14 @@ import {fetcher} from '@xmtp/proto';
 
 import config from '@unstoppabledomains/config';
 
-import type {SerializedAttachmentResponse} from '../../../lib';
-import {fetchApi} from '../../../lib';
 import {notifyEvent} from '../../../lib/error';
+import {fetchApi} from '../../../lib/fetchApi';
+import type {
+  SerializedAttachmentResponse} from '../../../lib/types/domain';
 import {
   getDomainSignatureExpiryKey,
   getDomainSignatureValueKey,
-} from '../../Wallet/ProfileManager';
+} from '../../../lib/types/domain';
 import {localStorageWrapper} from '../storage';
 
 export const uploadAttachment = async (
