+++ packages/ui-components/src/components/Wallet/AccessWallet.tsx
@@ -14,17 +14,17 @@ import React, {useEffect, useState} from 'react';
 import truncateEthAddress from 'truncate-eth-address';
 
 import AccessEthereum from '../../components/Wallet/AccessEthereum';
-import {useWeb3Context} from '../../hooks';
 import useFireblocksMessageSigner from '../../hooks/useFireblocksMessageSigner';
 import useFireblocksState from '../../hooks/useFireblocksState';
 import useFireblocksTxSigner from '../../hooks/useFireblocksTxSigner';
-import {WalletName} from '../../lib';
+import useWeb3Context from '../../hooks/useWeb3Context';
 import {
   ReactSigner,
   UD_COMPLETED_SIGNATURE,
 } from '../../lib/fireBlocks/reactSigner';
 import {getBootstrapState} from '../../lib/fireBlocks/storage/state';
 import useTranslationContext from '../../lib/i18n';
+import {WalletName} from '../../lib/types/wallet';
 import type {Web3Dependencies} from '../../lib/types/web3';
 import useAccessWalletStyles from '../../styles/components/accessWallet.styles';
 import {isEthAddress} from '../Chat/protocol/resolution';
