+++ packages/ui-components/src/components/Wallet/Header.tsx
@@ -439,7 +439,7 @@ export const Header: React.FC<Props> = ({
           {showMessages && (
             <UnstoppableMessaging
               address={address}
-              silentOnboard={true}
+              silentOnboard={!isChromeStorageSupported('local')}
               hideIcon={true}
               onPopoutClick={onMessagePopoutClick}
               disableSupportBubble
