+++ packages/ui-components/src/tests/mocks/wallet.ts
@@ -4,7 +4,7 @@ import type {TokenEntry} from '../../components/Wallet/Token';
 import type {
   SerializedPublicDomainProfileData,
   SerializedWalletBalance,
-} from '../../lib';
+} from '../../lib/types/domain';
 import {TokenType} from '../../lib/types/domain';
 import type {AccountAsset} from '../../lib/types/fireBlocks';
 import {VALID_ETH_ADDRESS} from '../common';
